split code into encoderUtil
